# Add HMAC init benchmark