--- a/tool/speed.cc
+++ b/tool/speed.cc
@@ -1161,6 +1161,25 @@ static bool SpeedHmac(const EVP_MD *md, const std::string &name,
   if (!selected.empty() && name.find(selected) == std::string::npos) {
     return true;
   }
+  TimeResults results;
+  const size_t key_len = EVP_MD_size(md);
+  std::unique_ptr<uint8_t[]> key(new uint8_t[key_len]);
+  BM_memset(key.get(), 0, key_len);
+
+  if (!TimeFunction(&results, [&]() -> bool {
+#if defined(OPENSSL_1_0_BENCHMARK)
+        BM_NAMESPACE::UniquePtr<HMAC_CTX> ctx(new HMAC_CTX);
+        HMAC_CTX_init(ctx.get());
+#else
+        BM_NAMESPACE::UniquePtr<HMAC_CTX> ctx(HMAC_CTX_new());
+#endif
+        return HMAC_Init_ex(ctx.get(), key.get(), key_len, md, NULL /* ENGINE */);
+      })) {
+    fprintf(stderr, "HMAC_Init_ex failed.\n");
+    ERR_print_errors_fp(stderr);
+    return false;
+  }
+  results.Print(name + " init");
 
   for (size_t chunk_len : g_chunk_lengths) {
     if (!SpeedHmacChunk(md, name, chunk_len)) {