### Description of changes: 
Add a benchmark that measures the time of HMAC init. This is similar to the other init benchmarks for ciphers.

Update AEAD init benchmark to reuse the same stack allocated AEAD context for each iteration. This removes the time for `EVP_AEAD_CTX_zero` and `EVP_AEAD_CTX_cleanup` from the benchmark loop. This was called automatically for the ScopedEVP_AEAD_CTX when the [stack object goes out of scope](https://github.com/aws/aws-lc/blob/main/include/openssl/aead.h#L487-L489).

The only other init benchmark for XTS [already reuses the same context for all iterations](https://github.com/aws/aws-lc/blob/main/tool/speed.cc#L1005-L1017). 

### Testing:
```
Did 1466000 HMAC-SHA256 init operations in 1000568us (1465167.8 ops/sec)
Did 1425000 HMAC-SHA256 (16 bytes) operations in 1000255us (1424636.7 ops/sec): 22.8 MB/s
Did 1255000 HMAC-SHA256 (256 bytes) operations in 1000347us (1254564.7 ops/sec): 321.2 MB/s
Did 807000 HMAC-SHA256 (1350 bytes) operations in 1001090us (806121.3 ops/sec): 1088.3 MB/s
Did 250000 HMAC-SHA256 (8192 bytes) operations in 1000884us (249779.2 ops/sec): 2046.2 MB/s
Did 137000 HMAC-SHA256 (16384 bytes) operations in 1000208us (136971.5 ops/sec): 2244.1 MB/s
```

AEAD init changes:
```
Before:
Did 12554000 AEAD-AES-128-GCM seal init operations in 1000004us (12553949.8 ops/sec)
Did 12578000 AEAD-AES-128-GCM seal init operations in 1000018us (12577773.6 ops/sec)
Did 12620500 AEAD-AES-128-GCM seal init operations in 1000019us (12620260.2 ops/sec)

After:
Did 12956500 AEAD-AES-128-GCM seal init operations in 1000002us (12956474.1 ops/sec)
Did 13118000 AEAD-AES-128-GCM seal init operations in 1000053us (13117304.8 ops/sec)
Did 13167250 AEAD-AES-128-GCM seal init operations in 1000006us (13167171.0 ops/sec)
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and the ISC license.